Deprecate old relayer code
